Implement premium account activation/deactivation and billing logic

Closes #11 and #73

See merge request informatik/qe/swapsws23/group8/g8t2!52

+++ src/main/java/at/qe/skeleton/internal/services/CreditCardService.java
@@ -22,6 +22,8 @@ public class CreditCardService {
 
   @Autowired private CreditCardRepository creditCardRepository;
 
+  @Autowired private UserxService userxService;
+
   public void saveCreditCard(CreditCard creditCard) throws IllegalArgumentException {
     if (!validateDate(creditCard.getExpirationDate())) {
       throw new IllegalArgumentException("Invalid expiration date.");
@@ -34,6 +36,13 @@ public class CreditCardService {
   }
 
   public void deleteCreditCard(CreditCard creditCard) {
+    if (creditCard == null) {
+      throw new IllegalArgumentException("Credit card cannot be null.");
+    }
+    if (userxService.isPremium(creditCard.getUser())) {
+      throw new IllegalArgumentException(
+          "You are a premium user. You cannot delete your credit card.");
+    }
     creditCardRepository.delete(creditCard);
   }
